test(python): split probe validation coverage
