+++ crates/sifr_diagnostics/src/codes/registry/registry_entries/python_interop.rs
@@ -47,7 +47,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Selected Python interpreter probe failed or returned invalid JSON.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_missing_interpreter_with_pyenv_0004",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_missing_interpreter_with_pyenv_0004",
         "Python environment probe failed: {reason}",
         "sifr_package::python",
         [arg!("reason"), json_arg!("interpreter"), json_arg!("venv")],
@@ -58,7 +58,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Selected Python interpreter is not supported CPython.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_non_cpython_json_with_pyenv_0005",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_non_cpython_json_with_pyenv_0005",
         "unsupported Python interpreter: {implementation}",
         "sifr_package::python",
         [arg!("implementation"), json_arg!("interpreter"), json_arg!("venv")],
@@ -69,7 +69,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Selected interpreter does not belong to the configured virtual environment.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_prefix_outside_venv_with_pyenv_0006",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_prefix_outside_venv_with_pyenv_0006",
         "Python sys.prefix is outside the selected virtual environment",
         "sifr_package::python",
         [
@@ -84,7 +84,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Selected Python environment has no site-packages path.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_missing_site_packages_with_pyenv_0007",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_missing_site_packages_with_pyenv_0007",
         "selected Python environment has no site-packages path",
         "sifr_package::python",
         [json_arg!("interpreter"), json_arg!("venv")],
@@ -95,7 +95,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Declared Python import root is missing from the selected environment.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_missing_declared_import_with_pyenv_0008",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_missing_declared_import_with_pyenv_0008",
         "declared Python import root is missing: {import_root}",
         "sifr_package::python",
         [arg!("import_root"), json_arg!("interpreter"), json_arg!("venv")],
@@ -106,7 +106,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Trusted native Python import root failed to load.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_native_import_failure_with_pyenv_0009",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_native_import_failure_with_pyenv_0009",
         "trusted native Python import root failed to load: {import_root}",
         "sifr_package::python",
         [
@@ -122,7 +122,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Free-threaded CPython is not supported for embedded interop.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_free_threaded_cpython_with_pyenv_0010",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_free_threaded_cpython_with_pyenv_0010",
         "free-threaded CPython is not supported",
         "sifr_package::python",
         [json_arg!("interpreter"), json_arg!("venv")],
@@ -133,7 +133,7 @@ pub(super) const ENTRIES: &[DiagnosticRegistryEntry] = &[
         "PYENV",
         "Configured Python project or lockfile is missing or stale.",
         Severity::Error,
-        "crates/sifr_package/src/python/tests.rs::probe_rejects_missing_lock_digest_with_pyenv_0011",
+        "crates/sifr_package/src/python/probe_validation_tests.rs::probe_rejects_missing_lock_digest_with_pyenv_0011",
         "Python environment metadata is stale: {reason}",
         "sifr_package::python",
         [arg!("reason"), json_arg!("interpreter"), json_arg!("venv")],
