+++ plans/issues/active/ad-hoc-declaration-first-python-interop.md
@@ -1879,6 +1879,12 @@ both one-sided doctor patches, deterministic byte-level non-mutation,
 multi-application coverage, focused test suites, and the maintained-source
 guardrails. The unrelated dirty Ruff submodule is absent from the PR diff.
 
+The post-review authoritative gate exposed the package-manager's stricter
+420-line module limit for `sifr_package` tests. Probe-validation cases now live
+in their own responsibility-focused module, diagnostic registry evidence paths
+were updated, and both the package-manager guardrail and all 133 package tests
+pass after the split.
+
 Tasks:
 
 - Add read-only `sifr python check` using the same package/driver plan as normal
